+++ src/util.rs
@@ -140,7 +140,7 @@ fn sockaddr_to_network_addr(sa: *const libc::sockaddr) -> (Option<MacAddr>, Opti
             (None, None)
         } else if (*sa).sa_family as libc::c_int == bpf::AF_LINK {
             let sdl: *const bpf::sockaddr_dl = mem::transmute(sa);
-            let nlen = (*sdl).sdl_nlen as uint;
+            let nlen = (*sdl).sdl_nlen as usize;
             let mac = MacAddr((*sdl).sdl_data[nlen + 0] as u8,
                               (*sdl).sdl_data[nlen + 1] as u8,
                               (*sdl).sdl_data[nlen + 2] as u8,
@@ -247,7 +247,7 @@ fn get_network_interfaces_impl() -> Vec<NetworkInterface> {
 
     let vec_size = adapters_size / mem::size_of::<winpcap::IP_ADAPTER_INFO>() as u32;
 
-    let mut adapters = Vec::with_capacity(vec_size as uint);
+    let mut adapters = Vec::with_capacity(vec_size as usize);
 
     // FIXME [windows] Check return code
     unsafe {
@@ -256,7 +256,7 @@ fn get_network_interfaces_impl() -> Vec<NetworkInterface> {
 
     // Create a complete list of NetworkInterfaces for the machine
     let mut cursor = adapters.as_mut_ptr();
-    let mut all_ifaces = Vec::with_capacity(vec_size as uint);
+    let mut all_ifaces = Vec::with_capacity(vec_size as usize);
     while cursor.is_not_null() {
         let mac = unsafe {
                     MacAddr((*cursor).Address[0],
